Тесты падают. Разбираться.

+++ Zzz/additional/experience/experience.cpp
@@ -11,7 +11,7 @@ int searchInsert(vector<int>& nums, int target)
 		size_t mid = left + (right - left) / 2;
 		if (nums[mid] == target)
 		{
-			return mid;
+			return (int)mid;
 		}
 		else if (nums[mid] < target)
 		{
